requirements.txt
# requirements.txt
tensorflow>=2.8.0
librosa>=0.9.0
soundfile>=0.10.0
numpy>=1.21.0
scikit-learn>=1.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
flask>=2.0.0
pydantic>=1.8.0
python-multipart>=0.0.5
noisereduce>=2.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.5.0
pyaudio>=0.2.11
websockets>=10.0
pyyaml>=6.0
pandas>=1.3.0